+++ src/app/js/model-extensions/model-sync-local.js
@@ -73,18 +73,18 @@ Object of key/value pairs to fall back on when localStorage is not available.
 @type Object
 @private
 **/
-LocalSync._data = {};
+
+LocalSync._data = LocalSync._data || {};
 
 /**
 Cache to quickly access a specific object with a given ID.
-This maps a model's ID to its reference inside of `LocalSync._data`.
 
-@property _idMap
-@type Object
+@property _store
+@type Array
 @private
 **/
 
-LocalSync._idMap = {};
+LocalSync._store = LocalSync._store || {};
 
 LocalSync.prototype = {
 
@@ -137,24 +137,14 @@ LocalSync.prototype = {
         // Pull in existing data from localStorage, if possible.
         // Otherwise, see if there's existing data on the local cache.
         if (store) {
-            try {
-                LocalSync._data[this.root] = Y.JSON.parse(store);
-            } catch (e) {
-                LocalSync._data[this.root] = [];
-            }
-        } else {
-            LocalSync._data[this.root] || (LocalSync._data[this.root] = []);
-        }
+            LocalSync._store[this.root] = store.split('|') || [];
 
-        // Map each model's ID to its reference inside of data, if there
-        // are already existing models inside of `localStorage`.
-        LocalSync._idMap[this.root] || (LocalSync._idMap[this.root] = {});
-        Y.Array.each(LocalSync._data[this.root], function (item) {
-            var id = item.id;
-            if (id) {
-                LocalSync._idMap[this.root][id] = item;
-            }
-        }, this);
+            Y.Array.each(LocalSync._store[this.root], function (id) {
+                LocalSync._data[id] = Y.JSON.parse(this.storage.getItem(id));
+            }, this);
+        } else {
+            LocalSync._store[this.root] || (LocalSync._store[this.root] = []);
+        }
     },
     
     // -- Public Methods -----------------------------------------------------------
@@ -174,7 +164,7 @@ LocalSync.prototype = {
       * **delete**: Delete an existing model.
 
     @param {Object} [options] Sync options
-    @param {callback} [callback] Called when the sync operation finishes.
+    @param {Function} [callback] Called when the sync operation finishes.
       @param {Error|null} callback.err If an error occurred, this parameter will
         contain the error. If the sync operation succeeded, _err_ will be
         falsey.
@@ -241,7 +231,12 @@ LocalSync.prototype = {
     @since 3.13.0
     **/
     _index: function () {
-        return LocalSync._data[this.root];
+        var store = LocalSync._store[this.root],
+            data  = Y.Array.map(store, function (id) {
+                return LocalSync._data[id];
+            });
+
+        return data;
     },
 
     /**
@@ -253,7 +248,7 @@ LocalSync.prototype = {
     @since 3.13.0
     **/
     _show: function () {
-        return LocalSync._idMap[this.root][this.get('id')] || null;
+        return LocalSync._data[this.get('id')] || null;
     },
     
     /**
@@ -265,13 +260,16 @@ LocalSync.prototype = {
     @since 3.13.0
     **/
     _create: function () {
-        var hash  = this.toJSON(),
-            data  = LocalSync._data[this.root],
-            idMap = LocalSync._idMap[this.root];
-        
+        var hash  = this.toJSON();
+
         hash.id = this.generateID(this.root);
-        data.push(hash);
-        idMap[hash.id] = hash;
+
+        LocalSync._data[hash.id] = hash;
+        if (this.storage) {
+            this.storage.setItem(hash.id, hash);
+        }
+
+        LocalSync._store[this.root].push(hash.id);
 
         this._save();
         return hash;
@@ -286,10 +284,21 @@ LocalSync.prototype = {
     @since 3.13.0
     **/
     _update: function () {
-        var hash = Y.merge(this.toJSON());
-        LocalSync._idMap[this.get('id')] = hash;
+        var hash = this.toJSON(),
+            id = this.get('id');
+
+        LocalSync._data[id] = hash;
         
+        if (this.storage) {
+            this.storage.setItem(id, hash);
+        }
+
+        if (Y.Array.indexOf(LocalSync._store[this.root], id) === -1) {
+            LocalSync._store[this.root].push(id);
+        }
+
         this._save();
+
         return hash;
     },
 
@@ -298,12 +307,27 @@ LocalSync.prototype = {
     from the in-memory object, and saves into localStorage if available.
     
     @method _destroy
-    @return {Object} The deleted object.
     @protected
     @since 3.13.0
     **/
     _destroy: function () {
-        delete LocalSync._idMap[this.get('id')];
+        var id = this.get('id'),
+            storage = this.storage;
+
+        if (!LocalSync._data[id]) {
+            return;
+        }
+
+        delete LocalSync._data[id];
+
+        if (storage) {
+            storage.removeItem(id);
+        }
+
+        LocalSync._store[this.root] = Y.Array.filter(LocalSync._store[this.root], function (item) {
+            return item.id != id;
+        });
+
         this._save();
         return this.toJSON();
     },
@@ -317,10 +341,10 @@ LocalSync.prototype = {
     @since 3.13.0
     **/
     _save: function () {
-        if (LocalSync._hasLocalStorage) {
-            this.storage && this.storage.setItem(
+        if (LocalSync._hasLocalStorage && this.storage) {
+            this.storage.setItem(
                 this.root,
-                Y.JSON.stringify(LocalSync._data[this.root])
+                LocalSync._store[this.root].join('|')
             );
         }
     }