+++ src/test/js/TestCase.js
@@ -73,7 +73,6 @@ YUITest.TestCase.prototype = {
      * Resumes a paused test and runs the given function.
      * @param {Function} segment (Optional) The function to run.
      *      If omitted, the test automatically passes.
-     * @return {Void}
      * @method resume
      */
     resume : function (segment) {
@@ -85,9 +84,8 @@ YUITest.TestCase.prototype = {
      * continue executing the given code.
      * @param {Function} segment (Optional) The function to run after the delay.
      *      If omitted, the TestRunner will wait until resume() is called.
-     * @param {int} delay (Optional) The number of milliseconds to wait before running
+     * @param {Number} delay (Optional) The number of milliseconds to wait before running
      *      the function. If omitted, defaults to `DEFAULT_WAIT` ms (10s).
-     * @return {Void}
      * @method wait
      */
     wait : function (segment, delay){
@@ -100,6 +98,40 @@ YUITest.TestCase.prototype = {
         }
 
         throw new YUITest.Wait(segment, delay);
+    },
+
+    /**
+    Creates a callback that automatically resumes the test. Parameters as passed
+    on to the callback.
+
+    @method next
+    @param {Function} callback Callback to call after resuming the test.
+    @return {Function} wrapped callback that resumes the test.
+    @example
+    ```
+    // using test.resume()
+    Y.jsonp(uri, function (response) {
+        test.resume(function () {
+            Y.Assert.isObject(response);
+        });
+    });
+    test.wait();
+
+    // using test.next()
+    Y.jsonp(uri, test.next(function (response) {
+        Y.Assert.isObject(response);
+    }));
+    test.wait();
+    ```
+    **/
+    next: function (callback) {
+        var self = this;
+        return function () {
+            var args = arguments;
+            self.resume(function () {
+                callback.apply(this, args);
+            });
+        };
     },
 
     /**
@@ -206,7 +238,6 @@ YUITest.TestCase.prototype = {
 
     /**
      * Function to run before each test is executed.
-     * @return {Void}
      * @method setUp
      */
     setUp : function () {
@@ -215,7 +246,6 @@ YUITest.TestCase.prototype = {
 
     /**
      * Function to run after each test is executed.
-     * @return {Void}
      * @method tearDown
      */
     tearDown: function () {