+++ src/test/js/TestRunner.js
@@ -34,7 +34,7 @@
         /**
          * A node in the test tree structure. May represent a TestSuite, TestCase, or
          * test function.
-         * @param {Variant} testObject A TestSuite, TestCase, or the name of a test function.
+         * @param {Any} testObject A TestSuite, TestCase, or the name of a test function.
          * @module test
  * @class TestNode
          * @constructor
@@ -44,7 +44,7 @@
 
             /**
              * The TestSuite, TestCase, or test function represented by this node.
-             * @type Variant
+             * @type {Any}
              * @property testObject
              */
             this.testObject = testObject;
@@ -100,8 +100,7 @@
             /**
              * Appends a new test object (TestSuite, TestCase, or test function name) as a child
              * of this node.
-             * @param {Variant} testObject A TestSuite, TestCase, or the name of a test function.
-             * @return {Void}
+             * @param {Any} testObject A TestSuite, TestCase, or the name of a test function.
              * @method appendChild
              */
             appendChild : function (testObject){
@@ -316,7 +315,6 @@
              * Adds a test case to the test tree as a child of the specified node.
              * @param {TestNode} parentNode The node to add the test case to as a child.
              * @param {Test.TestCase} testCase The test case to add.
-             * @return {Void}
              * @static
              * @private
              * @method _addTestCaseToTestTree
@@ -341,7 +339,6 @@
              * Adds a test suite to the test tree as a child of the specified node.
              * @param {TestNode} parentNode The node to add the test suite to as a child.
              * @param {Test.TestSuite} testSuite The test suite to add.
-             * @return {Void}
              * @static
              * @private
              * @method _addTestSuiteToTestTree
@@ -365,7 +362,6 @@
              * Builds the test tree based on items in the master suite. The tree is a hierarchical
              * representation of the test suites, test cases, and test functions. The resulting tree
              * is stored in _root and the pointer _cur is set to the root initially.
-             * @return {Void}
              * @static
              * @private
              * @method _buildTestTree
@@ -394,7 +390,6 @@
              * Handles the completion of a test object's tests. Tallies test results
              * from one level up to the next.
              * @param {TestNode} node The TestNode representing the test object.
-             * @return {Void}
              * @method _handleTestObjectComplete
              * @private
              */
@@ -744,7 +739,6 @@
              * to tell TestRunner about the error. This should never be called
              * by anyplace other than the Mock object.
              * @param {Error} error The error object.
-             * @return {Void}
              * @method _handleError
              * @private
              * @static
@@ -765,7 +759,6 @@
              * Runs a single test based on the data provided in the node.
              * @method _runTest
              * @param {TestNode} node The TestNode representing the test to run.
-             * @return {Void}
              * @static
              * @private
              */
@@ -837,7 +830,6 @@
              * The name assigned to the master suite of the TestRunner. This is the name
              * that is output as the root's name when results are retrieved.
              * @param {String} name The name of the result set.
-             * @return {Void}
              * @method setName
              */
             setName: function(name){
@@ -851,7 +843,6 @@
             /**
              * Adds a test suite or test case to the list of test objects to run.
              * @param testObject Either a TestCase or a TestSuite that should be run.
-             * @return {Void}
              * @method add
              * @static
              */
@@ -862,7 +853,6 @@
 
             /**
              * Removes all test objects from the runner.
-             * @return {Void}
              * @method clear
              * @static
              */
@@ -968,7 +958,6 @@
              * Resumes the TestRunner after wait() was called.
              * @param {Function} segment The function to run as the rest
              *      of the haulted test.
-             * @return {Void}
              * @method resume
              * @static
              */
@@ -986,7 +975,6 @@
              *      <code>oldMode</code> indicates the TestRunner should work in the YUI <= 2.8 way
              *      of internally managing test suites. <code>groups</code> is an array
              *      of test groups indicating which tests to run.
-             * @return {Void}
              * @method run
              * @static
              */