+++ src/yui/HISTORY.md
@@ -4,7 +4,10 @@ YUI Core Change History
 @VERSION@
 ------
 
-* No changes.
+* Fixed UA detection in recent versions of the Amazon Silk browser.
+  ([#1576][]: @adinardi)
+
+[#1576]: https://github.com/yui/yui3/pull/1576
 
 3.14.1
 ------