+++ src/yui/HISTORY.md
@@ -4,6 +4,7 @@ YUI Core Change History
 @VERSION@
 ------
 
+* Fixed value of `this` inside ES6 module definitions.
 * Fixed UA detection in recent versions of the Amazon Silk browser.
   ([#1576][]: @adinardi)