+++ src/yui/HISTORY.md
@@ -13,6 +13,11 @@ YUI().require('some-es6-module', function (Y, imports) {
 });
 ```
 
+* Fixed UA detection in recent versions of the Amazon Silk browser.
+  ([#1576][]: @adinardi)
+
+[#1576]: https://github.com/yui/yui3/pull/1576
+
 3.14.1
 ------